QUIK: Initrd support.
This adds the "old-kernel" option, which lets you boot passing
the arguments the old way supported by 2.4 and 2.2. It should
not be necessary, but who knows? Maybe some OF doesn't allow
you to set bootargs.

Of course, it means you can't use an initrd (I should probably
add params like initrd=0x%x@0x%x, in the hopes that support gets
anywhere for that).

Signed-off-by: Andrei Warkentin <awarkentin@vmware.com>