quik: all installing to hard drive without first stage
Infact this removes first stage support from the installer period.

By default the installer will install to the same hard drive that
second.b is on. It will install second.b into the Apple_Bootstrap
partition and Apple_Bootstrap partition only, making this
the bootable partition.

You can override the target device with '-d'. For example,
to install to the floppy drive (as a recovery aid) you could
do

$ quik -d /dev/fd0

Signed-off-by: Andrei Warkentin <awarkentin@vmware.com>