iquik: NewWorld changes
Make sure the user cannot hose their Yaboot installation
on their NewWorld by running iquik with default arguments.

The iquik installer should not be used on a NewWorld as
the bootblock booting method does not work on OF 3.0 systems.
However, someone could be using their NewWorld to create
media for an OldWorld...

Signed-off-by: Andrei Warkentin <andrey.warkentin@gmail.com>